[fix] gws/affair : download leave yearly (#4803)
* [fix] gws/affair : download leave yearly

* [fix] gws/affair : time_card download

* [add] spec
