+++ app/models/concerns/gws/addon/affair/leave_file.rb
@@ -1,7 +1,7 @@
 module Gws::Addon::Affair::LeaveFile
   extend ActiveSupport::Concern
   extend SS::Addon
-  include Gws::Affair::LeaveFile::Compensatory
+  include Compensatory
 
   included do
     attr_accessor :start_at_date, :start_at_hour, :start_at_minute,
