+++ app/models/gws/affair/capital_year/importer/day_count.rb
@@ -25,7 +25,7 @@ class Gws::Affair::CapitalYear::Importer::DayCount < Gws::Affair::CapitalYear::I
           count = setting.count
           leave_files = setting.annual_leave_files
           minutes = leave_files.map(&:leave_minutes_in_query).sum
-          effective_minutes = (setting.annual_leave_minutes - minutes)
+          effective_minutes = (setting.annual_leave_minutes(cur_site) - minutes)
           effective_minutes = effective_minutes > 0 ? effective_minutes : 0
 
           leaved_count = Gws::Affair::Utils.leave_minutes_to_day(cur_site, minutes)
